Playing with heuristics

+++ src/optimizer/Heuristic.java
@@ -2,5 +2,5 @@ package src.optimizer;
 
 
 public interface Heuristic<T extends Evolvable<T>> {
-	public Integer getValue(T evolvable);
+	public Double getValue(T evolvable);
 }
